requirements.txt
Django==5.0.2
djangorestframework==3.14.0
matplotlib==3.8.3
nltk==3.8.1
numpy==1.26.4
pillow==10.2.0
PyMuPDF==1.23.25
python-docx==1.1.0
wordcloud==1.9.3
